Fixed a couple compile issues I created in the last commit

+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICollectorMK1.java
@@ -34,7 +34,7 @@ public class GUICollectorMK1 extends GuiContainer
 	@Override
 	protected void drawGuiContainerForegroundLayer(int var1, int var2)
 	{
-		this.fontRenderer.drawString(Integer.toString(container.emc), 60, 32, 4210752);
+		this.fontRenderer.drawString(Long.toString(container.emc), 60, 32, 4210752);
 		
 		double kleinCharge = container.kleinEmc;
